* FIXED: some X64 warnings * MODIFIED: disabled the removal of player and pet auras on death due to it causing issues. FIXME: Auras need to be checked. * FIXED: packet sizes for refund packets * MODIFIED: being invincible will no longer cause mobs to ignore the Unit, and also it will actually ignore damage.

+++ src/arcemu-shared/NGLog.h
@@ -105,8 +105,8 @@ public:
     } 
 
     void dcds( char *str ){
-        int i = 0;
-        int len = strlen( str );
+        unsigned long i = 0;
+        size_t len = strlen( str );
         
         for(i = 0; i < len; ++i )
             str[i] = dcd( str[i] );
